feat: update stale fingerprint message
