+++ pkg/commands/ignore.go
@@ -7,16 +7,11 @@ import (
 
 	"github.com/bearer/bearer/pkg/flag"
 	"github.com/bearer/bearer/pkg/util/ignore"
-	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
 var migratedIgnoreComment = "migrated from bearer.yml"
-
-var bold = color.New(color.Bold).SprintFunc()
-var morePrefix = color.HiBlackString("├─ ")
-var lastPrefix = color.HiBlackString("└─ ")
 
 func NewIgnoreCommand() *cobra.Command {
 	usageTemplate := `
@@ -98,7 +93,7 @@ $ bearer ignore show <fingerprint>`,
 			if options.IgnoreShowOptions.All {
 				// show all fingerprints
 				for fingerprintId, fingerprint := range ignoredFingerprints {
-					cmd.Print(displayIgnoredEntryTextString(fingerprintId, fingerprint))
+					cmd.Print(ignore.DisplayIgnoredEntryTextString(fingerprintId, fingerprint))
 				}
 			} else {
 				// show a specific fingerprint
@@ -108,9 +103,9 @@ $ bearer ignore show <fingerprint>`,
 					cmd.Printf("Ignored fingerprint '%s' was not found in bearer.ignore file\n", fingerprintId)
 					return nil
 				}
-				cmd.Print(displayIgnoredEntryTextString(fingerprintId, selectedIgnoredFingerprint))
+				cmd.Print(ignore.DisplayIgnoredEntryTextString(fingerprintId, selectedIgnoredFingerprint))
 			}
-			cmd.Print("\n")
+			cmd.Print("\n\n")
 			return nil
 		},
 		SilenceErrors: false,
@@ -178,8 +173,8 @@ $ bearer ignore add <fingerprint> --author Mish --comment "Possible false positi
 			}
 
 			cmd.Print("Fingerprint added to bearer.ignore:\n\n")
-			cmd.Print(displayIgnoredEntryTextString(fingerprintId, ignoredFingerprints[fingerprintId]))
-			cmd.Print("\n")
+			cmd.Print(ignore.DisplayIgnoredEntryTextString(fingerprintId, ignoredFingerprints[fingerprintId]))
+			cmd.Print("\n\n")
 			return nil
 		},
 		SilenceErrors: false,
@@ -236,8 +231,8 @@ $ bearer ignore remove <fingerprint>`,
 			}
 
 			cmd.Print("Fingerprint successfully removed from bearer.ignore:\n\n")
-			cmd.Print(displayIgnoredEntryTextString(fingerprintId, removedFingerprint))
-			cmd.Print("\n")
+			cmd.Print(ignore.DisplayIgnoredEntryTextString(fingerprintId, removedFingerprint))
+			cmd.Print("\n\n")
 			return nil
 		},
 		SilenceErrors: false,
@@ -342,28 +337,4 @@ func getIgnoredFingerprintsFromConfig(configPath string) (ignoredFingerprintsFro
 	}
 
 	return ignoredFingerprintsFromConfig, nil
-}
-
-func displayIgnoredEntryTextString(fingerprintId string, entry ignore.IgnoredFingerprint) string {
-	prefix := morePrefix
-	result := fmt.Sprintf(bold(color.HiBlueString("%s \n")), fingerprintId)
-
-	if entry.Author == nil && entry.Comment == nil {
-		prefix = lastPrefix
-	}
-	result += fmt.Sprintf("%sIgnored At: %s\n", prefix, bold(entry.IgnoredAt))
-
-	if entry.Author != nil {
-		if entry.Comment == nil {
-			prefix = lastPrefix
-		}
-
-		result += fmt.Sprintf("%sAuthor: %s\n", prefix, bold(*entry.Author))
-	}
-
-	if entry.Comment != nil {
-		result += fmt.Sprintf("%sComment: %s\n", lastPrefix, bold(*entry.Comment))
-	}
-
-	return result
 }
